requirements.txt
fastapi
pyjwt
passlib[bcrypt]
asyncpg
redis
sqlalchemy
uvicorn
python-multipart 
bcrypt==3.2.0